Search endpoint

+++ lib/generate-embeddings.ts
@@ -373,6 +373,7 @@ export const generateFileEmbeddings = async (
   const { error } = await supabaseAdmin
     .from('file_sections')
     .insert(embeddingsData);
+
   if (error) {
     console.error('Error storing embeddings:', error);
     errors.push({
@@ -393,13 +394,10 @@ export const generateFileEmbeddings = async (
     );
   }
 
-  if (errors) {
+  if (errors?.length > 0) {
     // If there were errors, delete the file (which will cascade and delete
     // associated embeddings), to give a change to process the file again.
-    await supabaseAdmin
-      .from('file_sections')
-      .delete()
-      .filter('file_id', 'eq', fileId);
+    await supabaseAdmin.from('files').delete().eq('id', fileId);
   }
 
   return errors;
